Bug with resizing

+++ training_utils.py
@@ -954,7 +954,7 @@ def _process_images(raw_image, crop_size=640, scale_by=0.66, mu=127.0, scale=255
 
     # add a small amount of random noise to the size for variety
     nnoise = tf.random_normal([1], mean=1.0, stddev=0.025)
-    noisy_image_size = image_size * nnoise
+    noisy_image_size = tf.cast(image_size * nnoise, dtype=tf.int32)
 
     # random crop the image
     raw_image = tf.random_crop(raw_image, size=[noisy_image_size, noisy_image_size, 3])
